Create 200 * '<~>' + 50 * '     ' + 5 * '*****' + 5 * '     ' + 5 * '*****' + 110 * '     ' + 5 * '@@@@@' + 160 * '     ' + 3 * '[] [] [] [] [] [] []' + 13 * '     ' + 3 * '[] [] [] [] [] []'
Smiley face created from the Python command prompt  >>>
All code is on one line but I separated it so that its easier to read in github

NOTE:  The picture of a man with a goatee only appears in the minimized screen on my Windows computer.  You'll need to adjust the numbers for a full screen picture.